Rename subscriptions to feed and backend to data
