+++ src/data/BackendInterface.h
@@ -20,9 +20,9 @@
 #pragma once
 
 #include <ripple/ledger/ReadView.h>
-#include <backend/DBHelpers.h>
-#include <backend/SimpleCache.h>
-#include <backend/Types.h>
+#include <data/DBHelpers.h>
+#include <data/SimpleCache.h>
+#include <data/Types.h>
 #include <util/config/Config.h>
 #include <util/log/Logger.h>
