+++ src/data/SimpleCache.h
@@ -21,7 +21,7 @@
 
 #include <ripple/basics/base_uint.h>
 #include <ripple/basics/hardened_hash.h>
-#include <backend/Types.h>
+#include <data/Types.h>
 #include <map>
 #include <mutex>
 #include <shared_mutex>
