+++ src/etl/NFTHelpers.cpp
@@ -23,9 +23,9 @@
 #include <ripple/protocol/TxMeta.h>
 #include <vector>
 
-#include <backend/BackendInterface.h>
-#include <backend/DBHelpers.h>
-#include <backend/Types.h>
+#include <data/BackendInterface.h>
+#include <data/DBHelpers.h>
+#include <data/Types.h>
 
 namespace clio::etl {
 using namespace data;
