+++ src/rpc/RPC.cpp
@@ -45,7 +45,7 @@ Context::Context(
     uint32_t version_,
     boost::json::object const& params_,
     shared_ptr<BackendInterface const> const& backend_,
-    shared_ptr<subscription::SubscriptionManager> const& subscriptions_,
+    shared_ptr<feed::SubscriptionManager> const& subscriptions_,
     shared_ptr<etl::ETLLoadBalancer> const& balancer_,
     shared_ptr<etl::ReportingETL const> const& etl_,
     shared_ptr<web::WsBase> const& session_,
@@ -75,7 +75,7 @@ make_WsContext(
     boost::asio::yield_context& yc,
     boost::json::object const& request,
     shared_ptr<BackendInterface const> const& backend,
-    shared_ptr<subscription::SubscriptionManager> const& subscriptions,
+    shared_ptr<feed::SubscriptionManager> const& subscriptions,
     shared_ptr<etl::ETLLoadBalancer> const& balancer,
     shared_ptr<etl::ReportingETL const> const& etl,
     shared_ptr<web::WsBase> const& session,
@@ -116,7 +116,7 @@ make_HttpContext(
     boost::asio::yield_context& yc,
     boost::json::object const& request,
     shared_ptr<BackendInterface const> const& backend,
-    shared_ptr<subscription::SubscriptionManager> const& subscriptions,
+    shared_ptr<feed::SubscriptionManager> const& subscriptions,
     shared_ptr<etl::ETLLoadBalancer> const& balancer,
     shared_ptr<etl::ReportingETL const> const& etl,
     util::TagDecoratorFactory const& tagFactory,
