+++ src/rpc/RPCHelpers.cpp
@@ -18,7 +18,7 @@
 //==============================================================================
 
 #include <ripple/basics/StringUtilities.h>
-#include <backend/BackendInterface.h>
+#include <data/BackendInterface.h>
 #include <rpc/RPCHelpers.h>
 #include <util/Profiler.h>
 #include <util/log/Logger.h>
