+++ src/rpc/handlers/AccountChannels.cpp
@@ -25,8 +25,8 @@
 #include <ripple/protocol/jss.h>
 #include <boost/json.hpp>
 #include <algorithm>
-#include <backend/BackendInterface.h>
-#include <backend/DBHelpers.h>
+#include <data/BackendInterface.h>
+#include <data/DBHelpers.h>
 
 #include <rpc/RPCHelpers.h>
