+++ src/rpc/handlers/AccountCurrencies.cpp
@@ -26,7 +26,7 @@
 #include <boost/json.hpp>
 #include <algorithm>
 
-#include <backend/BackendInterface.h>
+#include <data/BackendInterface.h>
 #include <rpc/RPCHelpers.h>
 
 namespace clio::rpc {
