+++ src/rpc/handlers/AccountOffers.cpp
@@ -28,8 +28,8 @@
 #include <algorithm>
 #include <rpc/RPCHelpers.h>
 
-#include <backend/BackendInterface.h>
-#include <backend/DBHelpers.h>
+#include <data/BackendInterface.h>
+#include <data/DBHelpers.h>
 
 namespace clio::rpc {
