+++ src/rpc/handlers/LedgerData.cpp
@@ -17,7 +17,7 @@
 */
 //==============================================================================
 
-#include <backend/BackendInterface.h>
+#include <data/BackendInterface.h>
 #include <rpc/RPCHelpers.h>
 #include <util/log/Logger.h>
