+++ src/rpc/handlers/LedgerEntry.cpp
@@ -21,7 +21,7 @@
 #include <ripple/protocol/STLedgerEntry.h>
 #include <boost/json.hpp>
 
-#include <backend/BackendInterface.h>
+#include <data/BackendInterface.h>
 #include <rpc/RPCHelpers.h>
 // {
 //   ledger_hash : <ledger>
