+++ src/rpc/handlers/NFTInfo.cpp
@@ -21,7 +21,7 @@
 #include <ripple/protocol/Indexes.h>
 #include <boost/json.hpp>
 
-#include <backend/BackendInterface.h>
+#include <data/BackendInterface.h>
 #include <rpc/RPCHelpers.h>
 
 // {
