+++ src/rpc/handlers/ServerInfo.cpp
@@ -17,7 +17,7 @@
 */
 //==============================================================================
 
-#include <backend/BackendInterface.h>
+#include <data/BackendInterface.h>
 #include <etl/ETLSource.h>
 #include <etl/ReportingETL.h>
 #include <main/Build.h>
